refactor(contractpath): promote pathx to pkg/contractpath; drop kernel mirror
Supersedes the kernel/governance mirror introduced in the previous commit
on this branch (review feedback: a mirror plus a source-of-truth still leaves
a Soft drift surface across packages — see .claude/rules/gocell/ai-robust.md
§Hard 范本目录 "codegen funnel + golden" vs archtest 兜底).

Move tools/codegen/internal/pathx → pkg/contractpath so the contract ID →
generated package path mapping has one Go-level definition shared by all
consumers (kernel/governance CH-04/05/06, tools/codegen contractgen + cellgen,
archtest). kernel/ is permitted to import pkg/ (CLAUDE.md 分层依赖规则), so
the previous "kernel cannot import tools/codegen" justification for keeping
a mirror no longer applies once the helper lives under pkg/.

Changes:

- pkg/contractpath/internalapi.go: new package. Exposes
  ContractIDToPackagePath (joined "generated/contracts/..." string, used by
  contractgen/cellgen) and Segments (path-segment slice, used by
  kernel/governance with safeJoinUnderRoot for path-traversal protection).
  internalapi_test.go: ported from pathx + new TestSegmentsAgreesWithJoined
  which inverts Segments to ContractIDToPackagePath, locking the two
  callsite shapes to one rewrite rule.
- kernel/governance/rules_http.go: drop generatedContractPathSegments; both
  findCodegenHandlerFile (CH-04/05 handler lookup) and typedEnvelopeTypesGenPath
  (CH-06 typed envelope lookup) call contractpath.Segments directly.
- kernel/governance/rules_http_test.go: writeGeneratedHandlerFile helper uses
  contractpath.Segments. The two CH-04/05 internal-segment regression tests
  added in the previous commit are preserved (they validate end-to-end
  governance behavior, orthogonal to the source of the path mapping).
- tools/codegen/{contractgen,cellgen}/builder.go: import path switch
  tools/codegen/internal/pathx → pkg/contractpath.
- tools/codegen/contractgen/doc.go: cross-tooling reference updated.
- tools/archtest/adapter_returns_declared_types_test.go: comment refresh
  (importPathToContractID is the inverse of contractpath.ContractIDToPackagePath).
- tools/codegen/internal/pathx/: removed (single-source elimination).

ref: pkg/contractpath/internalapi.go godoc lists all four consumers.
